Move repo-docs scripts into portable skill bundle
