+++ package.json
@@ -22,7 +22,7 @@
     "release:checklist": "cat docs/release-checklist.md",
     "release:dry-run": "bash scripts/release-dry-run.sh",
     "docs:frontmatters": "node scripts/list-doc-frontmatters.mjs",
-    "docs:validate:test": "vitest run scripts/validate-doc-frontmatter.test.ts scripts/list-doc-frontmatters.test.ts",
+    "docs:validate:test": "vitest run .agents/skills/repo-docs/scripts/validate-doc-frontmatter.test.ts .agents/skills/repo-docs/scripts/list-doc-frontmatters.test.ts",
     "docs:validate": "node scripts/validate-doc-frontmatter.mjs",
     "typecheck": "tsc --noEmit",
     "test": "vitest run",
